Changed ordinal number from uint to byte. Added raw SQL query for getting the shopping list as LINQ doesn't want to do INNER JOIN with planned meals.

+++ src/Application/MediatR/PlannedMeal/Commands/CreatePlannedMealCommand.cs
@@ -7,5 +7,5 @@ using System.Threading.Tasks;
 
 namespace FoodPlanner.Application.MediatR.PlannedMeal.Commands
 {
-    public record CreatePlannedMealCommand(uint OrdinalNumber, DateTime ScheduledFor, int MealId) : IRequest<Domain.Entities.PlannedMeal>;
+    public record CreatePlannedMealCommand(byte OrdinalNumber, DateTime ScheduledFor, int MealId) : IRequest<Domain.Entities.PlannedMeal>;
 }
